% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_xlog.R
\name{clean_xlog}
\alias{clean_xlog}
\title{Clean extended log (xlog) files from NetHack 3.6.1.}
\usage{
clean_xlog(xlog)
}
\arguments{
\item{xlog}{An xlog file}
}
\value{
A data frame
}
\description{
Clean extended log (xlog) files from NetHack 3.6.1. Adds column names, removes unneeded columns and
converts variables to their correct format.
}
\examples{
\dontrun{
clean_xlog(xlog)
}
}
