% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_defeated_monsters.R
\name{extract_defeated_monsters}
\alias{extract_defeated_monsters}
\title{Extract information about defeated monsters from an xlogfile}
\usage{
extract_defeated_monsters(dumplog)
}
\arguments{
\item{xlog}{A raw xlogfile}
}
\value{
A data frame with information on vanquished, genocided and extincted monsters
}
\description{
Extract information about defeated monsters from an xlogfile
}
\examples{
\dontrun{
get_dumplog(xlog)
}
}
