% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dumplog.R
\name{get_dumplog}
\alias{get_dumplog}
\title{Get the dumplog of a NetHack 3.6.1 run}
\usage{
get_dumplog(xlog)
}
\arguments{
\item{xlog}{A data frame containing cleaned xlog data}
}
\value{
A data frame with the dumplog column added to the input data frame
}
\description{
Get the dumplog of a NetHack 3.6.1 run
}
\examples{
\dontrun{
get_dumplog(xlog)
}
}
